requirements.txt
pytest
pytest-cov
pytest-xdist
requests
jsonschema
python-dotenv